chore(constants): button 속성 상수 분리
